requirements.txt
alembic==1.6.5
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.2
dnspython==2.1.0
docopt==0.6.2
docx==0.2.4
docx-mailmerge==0.5.0
dominate==2.6.0
email-validator==1.1.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Mobility==1.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
freezegun==1.1.0
future==0.18.2
greenlet==1.1.0
html5==0.0.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
ntlm-auth==1.5.0
numpy>=1.22
openpyxl==3.0.7
pandas==1.1.5
#Pillow==8.3.2
pipreqs==0.4.10
platformdirs==2.2.0
pycparser==2.20
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SharePlum==0.5.1
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
virtualenv==20.7.0
visitor==0.1.3
#vsdx==0.4.15
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
yarg==0.1.9
xlrd==1.1.0
